Change code for render 1 meal to renderMeal.js and add a video feature
