+++ src/modules/api.js
@@ -8,7 +8,7 @@ const api = {
   },
   async getByIngredient() {
     const searchInputTxt = document.getElementById('search-input').value.trim();
-    const url1 = `https://www.themealdb.com/api/json/v1/1/filter.php?i=${searchInputTxt}`;
+    const url1 = `https://www.themealdb.com/api/json/v1/1/search.php?s=${searchInputTxt}`;
     const response = await fetch(url1);
     const x = await response.json();
     const ingredientMeal = await x.meals;
